Implement ranking support (#189)
Adds support for the qid parameter, allowing ranking to work correctly. The group parameter is not supported - this is also the case for XGBoost's Dask interface.

Support for LightGBM to come in the future. Not sure how that will look like, as it doesn't support the qid parameter.

Also fixes the ray_dmatrix_params arg being mandatory for RayXGBRanker.