Add `enable_categorical`, better detection for being in a PG (#235)
XGBoost>=1.5 added a new dmatrix param, `enable_categorical`. Adding it here.

This PR also makes the detection of the case of already being in a placement group more broad and not limited to just Tune. This allows for xgboost-ray to be ran in nested tasks that are not related to Tune (this should be two PRs really, but I needed one branch to patch it for hackathon).


Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>